analysis/requirements.txt
plotly>=5.18.0
pandas>=2.1.0
pyyaml>=6.0
python-frontmatter>=1.0.0
